+++ tests/integ/sagemaker/lineage/test_artifact.py
@@ -78,10 +78,12 @@ def test_list(artifact_objs, sagemaker_session):
 
 
 def test_downstream_trials(trial_associated_artifact, trial_obj, sagemaker_session):
-    # wait for TC to index
-    time.sleep(3)
-
-    trials = trial_associated_artifact.downstream_trials(sagemaker_session=sagemaker_session)
+    # allow trial components to index, 30 seconds max
+    for i in range(3):
+        time.sleep(10)
+        trials = trial_associated_artifact.downstream_trials(sagemaker_session=sagemaker_session)
+        if len(trials) > 0:
+            break
 
     assert len(trials) == 1
     assert trial_obj.trial_name in trials